+++ src/Root.xml
@@ -68,7 +68,7 @@
         -->
         <groups>
             <run>
-                <include name="test.signup" />
+                <include name="test.membership" />
             </run>
         </groups>
